requirements.txt
requests>=2.28.1
beautifulsoup4>=4.11.1
tqdm>=4.64.0
biopython>=1.79
openai>=1.3.0
python-dotenv>=1.0.0
matplotlib>=3.5.0
pandas>=1.3.0
numpy>=1.20.0
wordcloud>=1.8.0
scikit-learn>=1.0.0
